form is functional, just need to complete tests
